Attempt to make rescue mode more robust

+++ nixopshetzner/backends/server.py
@@ -360,24 +360,24 @@ class HetznerState(MachineState):
         self.partitions if available (which it shouldn't if you're not doing
         something nasty).
         """
-        self.log("rebooting machine ‘{0}’ ({1}) into rescue system"
-                 .format(self.name, self.main_ipv4))
         server = self._get_server_by_ip(self.main_ipv4)
-        server.rescue.activate()
-        rescue_passwd = server.rescue.password
-        self.log("rescue password is '{0}'".format(rescue_passwd))
-        if hard or (install and self.state not in (self.UP, self.RESCUE)):
-            self.log_start("sending hard reset to robot... ")
-            server.reboot('hard')
-        else:
-            self.log_start("sending reboot command... ")
-            if self.state == self.RESCUE:
-                self.run_command("(sleep 2; reboot) &", check=False)
+        if self.state != self.RESCUE or server.rescue.password is None:
+            self.log("rebooting machine ‘{0}’ ({1}) into rescue system"
+                     .format(self.name, self.main_ipv4))
+            server.rescue.activate()
+            if hard or (install and self.state not in (self.UP, self.RESCUE)):
+                self.log_start("sending hard reset to robot... ")
+                server.reboot('hard')
             else:
-                self.run_command("systemctl reboot", check=False)
-        self.log_end("done.")
-        self._wait_for_rescue(self.main_ipv4)
-        self.rescue_passwd = rescue_passwd
+                self.log_start("sending reboot command... ")
+                if self.state == self.RESCUE:
+                    self.run_command("(sleep 2; reboot) &", check=False)
+                else:
+                    self.run_command("systemctl reboot", check=False)
+            self.log_end("done.")
+            self._wait_for_rescue(self.main_ipv4)
+            self.rescue_passwd = server.rescue.password
+        self.log("rescue password is '{0}'".format(self.rescue_passwd))
         self.state = self.RESCUE
         self.ssh.reset()
         if bootstrap:
